+++ example/src/App.js
@@ -5,6 +5,7 @@ import { UseInput } from './UseInput'
 import { UseInterval } from './UseInterval'
 import { UseLocalStorage } from './UseLocalStorage'
 import { UseLockBodyScroll } from './UseLockBodyScroll'
+import { UseOnClickOutside } from './UseOnClickOutside'
 import { UseTimeLeft } from './UseTimeLeft'
 
 const App = () => {
@@ -16,6 +17,7 @@ const App = () => {
       <UseInterval />
       <UseTimeLeft />
       <UseLocalStorage />
+      <UseOnClickOutside />
       <UseLockBodyScroll />
     </>
   )